Removed custom Checker class and added joi.
The custom Checker class is made on the old fashioned way of doing
things, where i do things for myself, but in this project i like to
use a specialized library, because it is easy and provides more
flexibility since everything is ready and well-documented. Also this
library provides a lot of things that my old own class does not, so
it will help a lot in the speed of the development.